Combines the audit-fix branch with main's newer work (admin detail pages,
Android build/resource fixes, generation pipeline + Celery Beat fixes).

Conflicts resolved (only 3 files conflicted; 4 more auto-merged):

- app/middleware/rate_limit.py: keep main's richer implementation (per-endpoint
  limits, Redis sorted-set sliding window, Prometheus counters). It already
  falls back to the in-memory store when Redis fails, which is the fix the
  audit asked for, so there is no silent bypass.
- app/workers/generation_tasks.py: keep main's delegation to the real
  execute_pipeline task and re-apply the queue-pause check before hand-off, so
  the admin "pause queue" switch still defers work instead of consuming retries.
- pyproject.toml: union of both branches. Restore structlog (required by
  app/core/logging_config.py on main). Drop passlib, yookassa and pillow, which
  are unused - password hashing imports bcrypt directly and payments use raw
  webhooks.

Correctness fix for the merge: app/core/metrics.py is restored. It was deleted
as dead code on the audit branch, but main's exception_handlers.py and
rate_limit.py import http_requests_total from it, so deleting it would have
broken startup.

Verified on the audit branch before merging: app imports (210 paths / 244
operations), pytest 167 passed 1 skipped, tsc --noEmit 0 errors, ruff clean,
alembic single head 034 with migration 034 rendered offline to valid
PostgreSQL DDL (13 statements).

+++ backend/app/models/ab_test.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -48,10 +48,10 @@ class ABTestResult(Base, UUIDPrimaryKeyMixin, TimestampMixin):
     )
     variant_code: Mapped[str] = mapped_column(String(50), nullable=False)
     metric: Mapped[str] = mapped_column(String(100), nullable=False)
-    value: Mapped[float] = mapped_column(Integer, nullable=False)
+    value: Mapped[float] = mapped_column(Numeric, nullable=False)
     user_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    conversion_rate: Mapped[float | None] = mapped_column(Integer)
-    revenue_impact_rub: Mapped[float | None] = mapped_column(Integer)
+    conversion_rate: Mapped[float | None] = mapped_column(Numeric)
+    revenue_impact_rub: Mapped[float | None] = mapped_column(Numeric)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now(), nullable=False
     )
